requirements.txt
APScheduler==3.11.0
blinker==1.9.0
boto3==1.37.18
botocore==1.37.18
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
durationpy==0.9
Flask==3.1.0
Flask-APScheduler==1.13.1
google-analytics-data==0.18.18
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kubernetes==32.0.1
loguru==0.7.3
MarkupSafe==3.0.2
oauthlib==3.2.2
peewee==3.17.9
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.4
six==1.17.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
win32_setctime==1.2.0
